# swagger.yaml
openapi: 3.0.0
paths:
  /user/register:
    post:
      operationId: UsersController_createUser
      summary: Create a new user
      security:
        - access-token: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDto'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Bad Request - Caused by an incorrect request object
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Email must be a maximum of 50 characters
                        - Email is not valid
                        - Email cannot be empty
                        - Email must be a string
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.364Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.364Z'
                    description: Date of the error
        '409':
          description: Conflict - Caused by an email already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Email already exists
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 409
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.364Z'
                    description: Date of the error
      tags:
        - Users
...
  /parking:
    post:
      operationId: ParkingsController_createParking
      summary: Save parking
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingResponseDto'
        '400':
          description: Bad Request - Caused by an invalid user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - >-
                          User isn't a 'SOCIO', only users with role 'SOCIO' can
                          be assigned to the parking
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.354Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.354Z'
                    description: Date of the error
        '404':
          description: Not Found - Caused by user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User not found
                      description: List of errors
                  error:
                    type: string
                    example: EntityNotFoundException
                    description: Error type
                  status:
                    type: number
                    example: 404
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.354Z'
                    description: Date of the error
      tags: &ref_1
        - Parkings
      security: &ref_2
        - access-token: []
  /parking/{id}:
    put:
      operationId: ParkingsController_updateParking
      summary: Update parking
      parameters:
        - name: id
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingResponseDto'
        '400':
          description: Bad Request - Caused by an invalid user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - >-
                          User isn't a 'SOCIO', only users with role 'SOCIO' can
                          be assigned to the parking
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.355Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.355Z'
                    description: Date of the error
        '404':
          description: Not Found - Caused by user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User not found
                      description: List of errors
                  error:
                    type: string
                    example: EntityNotFoundException
                    description: Error type
                  status:
                    type: number
                    example: 404
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.355Z'
                    description: Date of the error
      tags: *ref_1
      security: *ref_2
    get:
      operationId: ParkingsController_getParkingById
      summary: Get parking by id
      parameters:
        - name: id
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingResponseDto'
        '400':
          description: Bad Request - Caused by parking not exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                      description: List of errors
                  error:
                    type: string
                    example: ParkingNotExistsException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.356Z'
                    description: Date of the error
      tags: *ref_1
      security: *ref_2
    delete:
      operationId: ParkingsController_deleteParking
      summary: Delete parking
      parameters:
        - name: id
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      responses:
        '200':
          description: Parking deleted
        '400':
          description: Bad Request - Caused by parking not exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                      description: List of errors
                  error:
                    type: string
                    example: ParkingNotExistsException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.356Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.356Z'
                    description: Date of the error
      tags: *ref_1
      security: *ref_2
  /parkings:
    get:
      operationId: ParkingsController_getAllParkings
      summary: Get all parking
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingResponseDto'
        '404':
          description: Not Found - Caused by not found any parking
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Not found any parking
                      description: List of errors
                  error:
                    type: string
                    example: NotFoundException
                    description: Error type
                  status:
                    type: number
                    example: 404
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.356Z'
                    description: Date of the error
      tags: *ref_1
      security: *ref_2
...
info:
  title: PARKING CRUD API
  description: >-
    The Rest API manages vehicle parking, allows partner creation for vehicle
    entry and exit, and uses JWT tokens with role-based access control (ADMIN
    and SOCIO) expiring every 6 hours, built with NestJS and PostgreSQL
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
  schemas:
    UserRequestDto:
      type: object
      properties:
        email:
          type: string
          description: User email
          maxLength: 50
          example: socio@mail.com
        password:
          type: string
          description: User password
          example: socio
      required:
        - email
        - password
    RoleDto:
      type: object
      properties:
        id:
          type: number
          description: Role id
          example: 2
        name:
          type: string
          description: Role name
          example: SOCIO
      required:
        - id
        - name
    UserResponseDto:
      type: object
      properties:
        id:
          type: number
          description: User id
          example: 2
        email:
          type: string
          description: User email
          example: socio@mail.com
        role:
          description: User role
          allOf:
            - $ref: '#/components/schemas/RoleDto'
      required:
        - id
        - email
        - role
...
    ParkingRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Parking name
          maxLength: 50
          example: Parking One
        hourlyCost:
          type: number
          description: Hourly cost
          example: 10.5
        maxParkingSpace:
          type: number
          description: Max parking space
          example: 100
        userId:
          type: number
          description: User id
          example: 2
      required:
        - name
        - hourlyCost
        - maxParkingSpace
        - userId
    ParkingResponseDto:
      type: object
      properties:
        id:
          type: number
          description: Parking id
          example: 1
        name:
          type: string
          description: Parking name
          example: Parking One
        hourlyCost:
          type: number
          description: Hourly cost
          example: 10.5
        maxParkingSpace:
          type: number
          description: Max parking space
          example: 100
        user:
          description: User
          allOf:
            - $ref: '#/components/schemas/UserResponseDto'
      required:
        - id
        - name
        - hourlyCost
        - maxParkingSpace
        - user
...